Initial port to 1.7.10

+++ xplat/src/main/java/dev/emi/emi/mixin/accessor/HandledScreenAccessor.java
@@ -1,11 +1,11 @@
 package dev.emi.emi.mixin.accessor;
 
+import net.minecraft.inventory.slot.Slot;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
-import net.minecraft.screen.slot.Slot;
 
 @Mixin(HandledScreen.class)
 public interface HandledScreenAccessor {
@@ -26,5 +26,5 @@ public interface HandledScreenAccessor {
 	int getBackgroundHeight();
 
 	@Invoker("getSlotAt")
-	Slot invokeGetSlotAt(double x, double y);
+	Slot invokeGetSlotAt(int x, int y);
 }
